requirements.txt
pyaudio
numpy
openai-whisper
requests
edge-tts
pygame
python-dotenv 